Add rules to build an rpm package
The rpm could be installed to an rpm based linux distribution.
Boot entries for android-x86 will be added. After installation,
reboot the system to android-x86.

To build the rpm, you need to install 'rpm-build' on a Fedora
based system, or 'rpm' on a Debian/Ubuntu based system.
Then do it by 'make rpm'.

Just install the rpm as a normal package on a Fedora based system.
On Debian/Ubuntu, you can use the 'alien' to install it:

$ sudo apt install alien
$ sudo alien -c -i android-x86-6.0-r1.i686.rpm