Add a more flexible android.cfg of grub-efi
The android.cfg is used in both installation media and the installed
system. So we don't need to replicate the code. More functions could
be added in the future.

If boot loaders of other OSes are found in ESP, add them to the boot
menu. Currently Fedora, Ubuntu and Windows are supported.

Also include custom.cfg if it exists.