1-install: allow to upgrade an arbitrary folder
Originally the installer only searches and upgrades known folders of
official releases. That causes the upgrade function is not so useful
for unofficial build. Besides, it's a burden to maintain the upgrade
list.

To improve that, search all folders containing android-x86 files.
If multiple folders are found, show a dialog to ask users which one
to upgrade.