init: do not keep initrd's busybox in debug mode
Keeping initrd's busybox in debug mode is an old feature before we added
busybox porting to Android-x86. But it introduced some inconsistencies
between debug and normal modes. Moreover, it depends on AOSP's tool (ln)
to work properly. In case the tool fails to run, the system may not boot.

In short, the feature does more harm than good. Let's remove it from
debug mode. Only keep it if BUSYBOX=1 is set in cmdline.